+++ compiler/libsolidity/ast/TypeProvider.cpp
@@ -1130,7 +1130,7 @@ std::array<std::unique_ptr<FixedBytesType>, 32> const TypeProvider::m_bytesM{{
 	{std::make_unique<FixedBytesType>(32)}
 }};
 
-std::array<std::unique_ptr<MagicType>, 8> const TypeProvider::m_magics{{
+std::array<std::unique_ptr<MagicType>, 9> const TypeProvider::m_magics{{
 	{std::make_unique<MagicType>(MagicType::Kind::Block)},
 	{std::make_unique<MagicType>(MagicType::Kind::Message)},
 	{std::make_unique<MagicType>(MagicType::Kind::Transaction)},
@@ -1138,7 +1138,8 @@ std::array<std::unique_ptr<MagicType>, 8> const TypeProvider::m_magics{{
 	{std::make_unique<MagicType>(MagicType::Kind::TVM)},
 	{std::make_unique<MagicType>(MagicType::Kind::Math)},
 	{std::make_unique<MagicType>(MagicType::Kind::Rnd)},
-	{std::make_unique<MagicType>(MagicType::Kind::Gosh)}
+	{std::make_unique<MagicType>(MagicType::Kind::Gosh)},
+	{std::make_unique<MagicType>(MagicType::Kind::BLS)}
 	// MetaType is stored separately
 }};