+++ compiler/libsolidity/ast/Types.cpp
@@ -3205,6 +3205,30 @@ std::string FunctionType::richIdentifier() const
 	std::string id = "t_function_";
 	switch (m_kind)
 	{
+	case Kind::BlsVerify: id += "blsverify"; break;
+	case Kind::BlsAggregate: id += "blsaggregate"; break;
+	case Kind::BlsFastAggregateVerify: id += "blsfastaggregateverify"; break;
+	case Kind::BlsAggregateVerify: id += "blsaggregateverify"; break;
+	case Kind::BlsG1Add: id += "blsg1add"; break;
+	case Kind::BlsG1Sub: id += "blsg1sub"; break;
+	case Kind::BlsG1Neg: id += "blsg1neg"; break;
+	case Kind::BlsG1Mul: id += "blsg1mul"; break;
+	case Kind::BlsMapToG1: id += "blsmaptog1"; break;
+	case Kind::BlsG1IsZero: id += "blsg1iszero"; break;
+	case Kind::BlsG1InGroup: id += "blsg1ingroup"; break;
+	case Kind::BlsG2Add: id += "blsg2add"; break;
+	case Kind::BlsG2Sub: id += "blsg2sub"; break;
+	case Kind::BlsG2Neg: id += "blsg2neg"; break;
+	case Kind::BlsG2Mul: id += "blsg2mul"; break;
+	case Kind::BlsMapToG2: id += "blsmaptog2"; break;
+	case Kind::BlsG2IsZero: id += "blsg2iszero"; break;
+	case Kind::BlsG2InGroup: id += "blsg2ingroup"; break;
+	case Kind::BlsG1Zero: id += "blsg1zero"; break;
+	case Kind::BlsG2Zero: id += "blsg2zero"; break;
+	case Kind::BlsPushR: id += "blspushr"; break;
+	case Kind::BlsG1MultiExp: id += "blsg1multiexp"; break;
+	case Kind::BlsG2MultiExp: id += "blsg2multiexp"; break;
+
 	case Kind::IntCast: id += "integercast"; break;
 
 	case Kind::StructUnpack: id += "structunpack"; break;
@@ -3349,6 +3373,7 @@ std::string FunctionType::richIdentifier() const
 	case Kind::MathMulDivMod: id += "mathmuldivmod"; break;
 	case Kind::MathDivMod: id += "mathdivmod"; break;
 	case Kind::MathSign: id += "mathsign"; break;
+	case Kind::MathMulMod: id += "mathmulmod"; break;
 
 	case Kind::MappingAt: id += "mappingat"; break;
 	case Kind::MappingDelMinOrMax: id += "mapdelmin"; break;
@@ -4533,6 +4558,8 @@ std::string MagicType::richIdentifier() const
 		return "t_magic_rnd";
 	case Kind::Gosh:
 		return "t_magic_gosh";
+	case Kind::BLS:
+		return "t_magic_bls";
 	}
 	return "";
 }
@@ -4841,6 +4868,12 @@ MemberList::MemberMap MagicType::nativeMembers(ASTNode const*) const
 				TypeProvider::function(
 					{}, {}, {}, {}, FunctionType::Kind::MathDivR, StateMutability::Pure
 				)
+			},
+			{
+				"mulmod",
+				TypeProvider::function(
+					{}, {}, {}, {}, FunctionType::Kind::MathMulMod, StateMutability::Pure
+				)
 			}
 		};
 		members.emplace_back("max", TypeProvider::function(
@@ -5132,6 +5165,263 @@ MemberList::MemberMap MagicType::nativeMembers(ASTNode const*) const
 				nullptr, FunctionType::Options::withArbitraryParameters()
 			)}
 		});
+	case Kind::BLS: {
+		return MemberList::MemberMap({
+			{
+				"verify",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice(), TypeProvider::tvmslice(), TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::boolean()},
+					strings{"", "", ""},
+					strings{""},
+					FunctionType::Kind::BlsVerify,
+					StateMutability::Pure
+				)
+			},
+			{
+				"aggregate",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmVector(TypeProvider::tvmslice())},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsAggregate,
+					StateMutability::Pure
+				)
+			},
+			{
+				"fastAggregateVerify",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmVector(TypeProvider::tvmslice()), TypeProvider::tvmslice(), TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::boolean()},
+					strings{"", "", ""},
+					strings{""},
+					FunctionType::Kind::BlsFastAggregateVerify,
+					StateMutability::Pure
+				)
+			},
+			{
+				"aggregateVerify",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmVector(TypeProvider::tuple({TypeProvider::tvmslice(), TypeProvider::tvmslice()})), TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::boolean()},
+					strings{"", ""},
+					strings{""},
+					FunctionType::Kind::BlsAggregateVerify,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1Add",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice(), TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{"", ""},
+					strings{""},
+					FunctionType::Kind::BlsG1Add,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1Sub",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice(), TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{"", ""},
+					strings{""},
+					FunctionType::Kind::BlsG1Sub,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1Neg",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG1Neg,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1Mul",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice(), TypeProvider::int257()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{"", ""},
+					strings{""},
+					FunctionType::Kind::BlsG1Mul,
+					StateMutability::Pure
+				)
+			},
+			{
+				"mapToG1",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsMapToG1,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1IsZero",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::boolean()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG1IsZero,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1InGroup",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::boolean()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG1InGroup,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2Add",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice(), TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{"", ""},
+					strings{""},
+					FunctionType::Kind::BlsG2Add,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2Sub",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice(), TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{"", ""},
+					strings{""},
+					FunctionType::Kind::BlsG2Sub,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2Neg",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG2Neg,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2Mul",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice(), TypeProvider::int257()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{"", ""},
+					strings{""},
+					FunctionType::Kind::BlsG2Mul,
+					StateMutability::Pure
+				)
+			},
+			{
+				"mapToG2",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsMapToG2,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2IsZero",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::boolean()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG2IsZero,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2InGroup",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmslice()},
+					TypePointers{TypeProvider::boolean()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG2InGroup,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1Zero",
+				TypeProvider::function(
+					TypePointers{},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{},
+					strings{""},
+					FunctionType::Kind::BlsG1Zero,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2Zero",
+				TypeProvider::function(
+					TypePointers{},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{},
+					strings{""},
+					FunctionType::Kind::BlsG2Zero,
+					StateMutability::Pure
+				)
+			},
+			{
+				"r",
+				TypeProvider::function(
+					TypePointers{},
+					TypePointers{TypeProvider::uint(255)},
+					strings{},
+					strings{""},
+					FunctionType::Kind::BlsPushR,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g1MultiExp",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmVector(TypeProvider::tuple({TypeProvider::tvmslice(),TypeProvider::int257()}))},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG1MultiExp,
+					StateMutability::Pure
+				)
+			},
+			{
+				"g2MultiExp",
+				TypeProvider::function(
+					TypePointers{TypeProvider::tvmVector(TypeProvider::tuple({TypeProvider::tvmslice(),TypeProvider::int257()}))},
+					TypePointers{TypeProvider::tvmslice()},
+					strings{""},
+					strings{""},
+					FunctionType::Kind::BlsG2MultiExp,
+					StateMutability::Pure
+				)
+			}
+		});
+	}
 	case Kind::Gosh: {
 		MemberList::MemberMap members;
 		for (auto const&[name, type] : std::vector<std::tuple<std::string, FunctionType::Kind>>{
@@ -5300,6 +5590,8 @@ std::string MagicType::toString(bool _withoutDataLocation) const
 		return "rnd";
 	case Kind::Gosh:
 		return "gosh";
+	case Kind::BLS:
+		return "bls";
 	}
 	solAssert(false, "Unknown kind of magic.");
 	return {};
@@ -6023,6 +6315,26 @@ MemberList::MemberMap TvmSliceType::nativeMembers(ASTNode const *) const {
 			)
 		},
 		{
+			"bitEmpty", TypeProvider::function(
+				{},
+				{TypeProvider::boolean()},
+				{},
+				{{}},
+				FunctionType::Kind::TVMSliceEmpty,
+				StateMutability::Pure
+			)
+		},
+		{
+			"refEmpty", TypeProvider::function(
+				{},
+				{TypeProvider::boolean()},
+				{},
+				{{}},
+				FunctionType::Kind::TVMSliceEmpty,
+				StateMutability::Pure
+			)
+		},
+		{
 			"bits", TypeProvider::function(
 				TypePointers{},
 				TypePointers{TypeProvider::uint(10)},
@@ -6101,7 +6413,27 @@ MemberList::MemberMap TvmSliceType::nativeMembers(ASTNode const *) const {
 				FunctionType::Kind::TVMSliceCompare,
 				StateMutability::Pure
 			)
-		}
+		},
+		{
+			"startsWithOne", TypeProvider::function(
+				TypePointers{},
+				TypePointers{TypeProvider::boolean()},
+				strings{},
+				strings{""},
+				FunctionType::Kind::TVMSliceSize,
+				StateMutability::Pure
+			)
+		},
+		{
+			"startsWith", TypeProvider::function(
+				TypePointers{TypeProvider::tvmslice()},
+				TypePointers{TypeProvider::boolean()},
+				strings{""},
+				strings{""},
+				FunctionType::Kind::TVMSliceSize,
+				StateMutability::Pure
+			)
+		},
 	};
 	return members;
 }
@@ -6225,14 +6557,26 @@ TypeResult TvmVectorType::unaryOperatorResult(Token _operator) const {
 
 MemberList::MemberMap TvmVectorType::nativeMembers(const ASTNode *) const
 {
+	TypePointers comps;
+	strings names;
+	if (auto tuple = dynamic_cast<TupleType const*>(valueType())) {
+		for (Type const* comp : tuple->components()) {
+			comps.emplace_back(comp);
+			names.emplace_back("");
+		}
+	} else {
+		comps.emplace_back(valueType());
+		names.emplace_back("");
+	}
+
 	MemberList::MemberMap members =
 	{
 		{
 			"push",
 			TypeProvider::function(
-				TypePointers{valueType()},
+				comps,
 				TypePointers{},
-				strings{std::string()},
+				names,
 				strings{},
 				FunctionType::Kind::TVMVectorPush,
 				StateMutability::Pure
@@ -6253,9 +6597,9 @@ MemberList::MemberMap TvmVectorType::nativeMembers(const ASTNode *) const
 			"pop",
 			TypeProvider::function(
 				TypePointers{},
-				TypePointers{valueType()},
+				comps,
 				strings{},
-				strings{std::string("last")},
+				names,
 				FunctionType::Kind::TVMVectorPop,
 				StateMutability::Pure
 			)
@@ -6275,9 +6619,9 @@ MemberList::MemberMap TvmVectorType::nativeMembers(const ASTNode *) const
 			"last",
 			TypeProvider::function(
 				TypePointers{},
-				TypePointers{valueType()},
+				comps,
 				strings{},
-				strings{std::string("last")},
+				names,
 				FunctionType::Kind::TVMVectorLast,
 				StateMutability::Pure
 			)